requirements.txt
asyncpg
httpx
openai
pydantic
python-dotenv
rich
typer
pytest
pytest-asyncio
